Feature/skip type creation (#18)
Implements option to skip the creation of types in the generated code via --skip-type-creation. This can be useful when we want to use our typescript code as source of truth for the types (often containing comments which we would otherwise get rid of).
Has to be used in combination with transformValue in order to work. Otherwise typescript would (currently, version 5.0.4) error out when trying to import a type and a value with the same name from two different files.